+++ enlighten/common.py
@@ -13,7 +13,7 @@ application.
       can be modules (files) within it
 """
 
-VERSION = "4.1.11"
+VERSION = "4.1.15"
 
 ctl = None