+++ enlighten/measurement/Measurement.py
@@ -243,6 +243,7 @@ class Measurement:
                            'Prefix',
                            'Suffix',
                            'Preset',
+                           'Auto-Raman',
                            'Session Count',
                            'Plugin Name']
 
@@ -833,15 +834,15 @@ class Measurement:
         if (self.processed_reading.reading and
             self.processed_reading.reading.take_one_request and
             self.processed_reading.reading.take_one_request.auto_raman_request):
-            log.debug("get_metadata: using Auto-Raman settings")
+            if field == "auto-raman":            return True
             if field == "integration time":      return self.processed_reading.reading.new_integration_time_ms
             if field == "scan averaging":        return self.processed_reading.reading.sum_count
             if field == "ccd gain":              return self.processed_reading.reading.new_gain_db if self.settings.is_sig() else self.settings.eeprom.detector_gain
-           
-        log.debug("get_metadata: not using Auto-Raman settings")
-        if field == "integration time":          return self.settings.state.integration_time_ms
-        if field == "scan averaging":            return self.settings.state.scans_to_average
-        if field == "ccd gain":                  return self.settings.state.gain_db if self.settings.is_sig() else self.settings.eeprom.detector_gain
+        else: 
+            if field == "auto-raman":            return False
+            if field == "integration time":      return self.settings.state.integration_time_ms
+            if field == "scan averaging":        return self.settings.state.scans_to_average
+            if field == "ccd gain":              return self.settings.state.gain_db if self.settings.is_sig() else self.settings.eeprom.detector_gain
 
         if field == "enlighten version":         return common.VERSION
         if field == "measurement id":            return self.measurement_id